CLDR-11681 Tailor currency format/symbol for KRW and JPY in ko and ja
In ko locale, use '#,##0.00¤' for KRW with '원' as the symbol.
In ja locale, do the same for JPY with '円' as the symbol.
In ko-KP locale, do the same for KPW while preserving the current
format for KRW.

Shorten the display names for KRW, KRH, JPY, KPW by dropping the
official (and long in case KPW) country names or replacing
it with colloquial short names.