CLDR-16433 Clean up language around spaceReplacement (#2768)
* CLDR-16433 Clean up language around spaceReplacement

* CLDR-16433 Refine the script determination

Rich, I tried new language for the script determination, to incorporate your method but added a bit to catch edge cases. See what you think.

* CLDR-16433 Fix hanging chad

* CLDR-16433 Revise as per discussion

* CLDR-16433 Another cleanup pass

* CLDR-16433 fix typos, add empty surname handling

* CLDR-16433 Clean up TOC, add handling core and prefix

I realized that there wasn't a discussion of the core and prefix, so added another section. Also reorganized the structure a bit so that the flow was top to bottom, and fixed the TOC.

* CLDR-16433 More cleanup

Have to commit to see the rich text diff, so sometimes takes multiple commits for cleanup.

* CLDR-16433 fix wording

Fix for https://github.com/unicode-org/cldr/pull/2768#discussion_r1131703276