CLDR-16717 Fix six more Persian to IPA transform issues (#3091)
* CLDR-16717 Fix Persian to IPA transform for <yeh hamza, sukun>

Since NFD reorders sukun before hamza, we need to have a special
line to take care of <yeh hamza, sukun> which is normalized to
<yeh, sukun, hamza>.

* CLDR-16717 Fix Persian to IPA transform for <yeh hamza, kasra>

Since NFD transforms <yeh hamza, kasra> to <yeh, kasra, hamza>, we
need to have a special rule to take care of its transformation.

* CLDR-16717 Fix Persian to IPA transform of <alef, waw> and <khah, waw, yeh>

<alef, waw> was transformed to /uː/ which is only correct at the beginning
of words: mid-word, it should be transformed to /ɒːv/.

In <khah, waw, yeh>, a relatively rare combination, the waw is no longer
pronounced in modern Persian, so it should be transformed to /χiː/.

* CLDR-16717 Fix Persian to IPA transform for various waw forms

The previous rules removed waw in some cases which was a hack.
This replaces them with various rules that take care of the missing
waw cases instead. Various test data that included a wrong extra
damma are also fixed to remove the damma.

* CLDR-16717 Fix Persian to IPA transform for <farsi yeh, vowel> cases

This fixes various cases in the transform where farsi yeh is
followed by a vowel.