CLDR-7405 Forgotten Password flow
- remove some dead code from LoginButton
- add a new ResetPassword and CldrLogin panels
- add a new option to the KeepLoggedInManager to generate a 1-hour token for use in email

At present the flow is that if you forgot your password, you will be given an option
to send a link in the mail. For smoketest, the link shows up on the console for testing.

Currently, the link takes you to a panel where you can remain logged into the Survey Tool.
In the future we can add logic to allow changing the password from there.