CLDR-17897 Make ConvertLanguageData Consistent (#4015)
If we re-run ConvertLanguageData on unrelated data, it will update the order and values of some other data -- this fixes inconsistencies with the XML outputs to match expectations. The biggest change was updating values in `language_script.tsv` to demote script variations to secondary when they really are not expected. Furthermore I added explicit annotations to `country_language_population.tsv` when the writing system for a country was a variant.

Scripts ran:

 mvn package -DskipTests=true
 java -jar tools/cldr-code/target/cldr-code.jar ConvertLanguageData
 java -jar tools/cldr-code/target/cldr-code.jar GenerateLikelySubtags