CLDR-15391 GenerateLikelySubtags cleanup
Removed hardcoded _CA entries because they aren't necessary since the likely subtags can be derived from the population data.

Also edited the cr_Cans_CA comment because it was overflowing the line and to add more context.

CLDR-15931 Restore und_Cans -> iu_Cans_CA

After discussing with the DDL committee, especially with someone who works with the Cree community -- we decided it was best to treat the Cree languages as separate languages and not a macrolanguage. For CLDR's purposes, [cr] is an alias of [cwd] / Woods Cree, so the likely subtag should not be cr_Cans_CA.