CLDR-18349 Keep scriptData when generating ConvertLanguageData (#4411)
ConvertLanguageData rebuilds part of the supplementalData.xml. Since the new is in the middle of 2 generated sections we need to re-write the generator so it does not accidentally drop it when running the script. Long term this is only more reason to break up the supplementalData.xml - but let's plan that out first.

Code to run generator:
```
mvn package -DskipTests=true &&  java -jar tools/cldr-code/target/cldr-code.jar ConvertLanguageData
```