CLDR-16715 Match Classical Chinese with Traditional Han writing
Classical Chinese [lzh] is the standardized form of Chinese that arose around the start of the Common Era (potentially as early as 5th Century BCE). It is marked as ending in common usage by the May Fourth Movement in 1919 -- it was replaced with vernacular Chinese [zh & it's constituents].

It's entire duration coincides with Traditional Han writing [Hant]. A few early writings were in the earlier scripts, but the vast majority of corpora are in Traditional script. The ending of common usage was 37 year before the advent of Simplified Chinese (1956). Since, for historic languages, we match their writing system to the largest corpora, we should fix this and match it to Traditional Han writing.