CLDR-18461 Update ISO languages for v48
I was building a new dataset with ISO codes and I thought "I should check to make sure the CLDR ISO codes are up to date". Turns out there are a few changes.

Followed the protocol in this document:

https://cldr.unicode.org/development/updating-codes/update-languagescriptregion-subtags

Question: @btangmu's previous PR added the languages to the XMLs but it doesn't look like the script did that for him -- do I have to do that manually? https://github.com/unicode-org/cldr/pull/3700