CLDR-19013 Update Gondi writing systems
Gunjala Gondi was recently revived and its primary used by the wsg (Adilabad Gondi) community -- but it's not widely used. Rather, Telugu is still the major script.

This change adds additional data for the Gondi languages in `langauges_script.tsv` as well as a population entry in `country_language_population.tsv` so the `LikelySubtags.xml` entry can be updated.